+++ src/compiler/copying.c
@@ -294,6 +294,9 @@ Expr *copy_expr(CopyStruct *c, Expr *source_expr)
 		case EXPR_OTHER_CONTEXT:
 			MACRO_COPY_EXPR(expr->expr_other_context.inner);
 			return expr;
+		case EXPR_NAMED_ARGUMENT:
+			MACRO_COPY_EXPR(expr->named_argument_expr.value);
+			return expr;
 		case EXPR_EMBED:
 			MACRO_COPY_EXPR(expr->embed_expr.len);
 			MACRO_COPY_EXPR(expr->embed_expr.filename);